safari-534.53-branch/LayoutTests/svg/dynamic-updates/SVGFEGaussianBlurElement-svgdom-in-prop-expected.txt

CONSOLE MESSAGE: line 40: TypeError: 'undefined' is not an object (evaluating 'blurElement.in1.baseVal = "SourceGraphic"')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' property of the SVGFEGaussianBlurElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL blurElement.in1.baseVal should be SourceAlpha. Threw exception TypeError: 'undefined' is not an object (evaluating 'blurElement.in1.baseVal')
